// apps/web/src/api-client/model/adminReportListItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Waymate API
 * Carpooling backend powering the Waymate web app.
 * OpenAPI spec version: 0.1.0
 */
import type { ReportId } from "./reportId";
import type { ReportType } from "./reportType";
import type { ReportStatus } from "./reportStatus";
import type { AdminReportListItemRideId } from "./adminReportListItemRideId";
import type { AdminReportListItemReporter } from "./adminReportListItemReporter";
import type { AdminReportListItemTarget } from "./adminReportListItemTarget";

export interface AdminReportListItem {
    id: ReportId;
    reportType: ReportType;
    reportStatus: ReportStatus;
    description: string;
    /** @nullable */
    rideId: AdminReportListItemRideId;
    reporter: AdminReportListItemReporter;
    target: AdminReportListItemTarget;
    createdAt: string;
}